pyproject.toml
[project]
name = "uc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.190.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
]
